// apps/head-office/src/components/ReportsSection/types.ts
export interface ITableHeadData {
  field: string;
  header: string;
  isRemovable?: boolean;
}

export interface Representative {
    name: string;
    image: string;
}
export interface Country {
    name: string;
    code: string;
}
export interface Customer {
    id: number;
    name: string;
    country: Country;
    company: string;
    date: string;
    status: string;
    verified: boolean;
    activity: number;
    representative: Representative;
    balance: number;
}